feat: confirmation dialog

+++ src/app/home/home.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
 import { Router } from '@angular/router';
 import { FormService } from '../services/form.service';
 import { FormRecord } from '../models/formRecord.interface';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-home',
@@ -15,7 +17,8 @@ export class HomeComponent implements OnInit {
   constructor(
     private router: Router,
     private formService: FormService,
-    private cdr: ChangeDetectorRef
+    private cdr: ChangeDetectorRef,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -32,9 +35,15 @@ export class HomeComponent implements OnInit {
   }
 
   deleteRecord(record: FormRecord) {
-    this.formService.deleteRecord(record.no);
-    this.records = [...this.formService.getRecords()];
-    this.cdr.detectChanges();
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent);
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.formService.deleteRecord(record.no);
+        this.records = [...this.formService.getRecords()];
+        this.cdr.detectChanges();
+      }
+    });
   }
 
   refreshRecords() {
